a2oa-server_tolower.sh

+++ games-server/a2oa-server/files/a2oa-server_tolower.sh
@@ -4,10 +4,10 @@ IFS=$'\n'
 
 
 
-SERVERPATH=/opt/a2oa-server/
+SERVERPATH=${SERVERPATH:-"/opt/a2oa-server/"}
 
-ARMA2STEAMPATH="/media/s3/games/SteamLibraryProton/steamapps/common/Arma 2"
-ARMA2OASTEAMPATH="/media/s3/games/SteamLibraryProton/steamapps/common/Arma 2 Operation Arrowhead"
+ARMA2STEAMPATH=${ARMA2STEAMPATH:-"/media/s3/games/SteamLibraryProton/steamapps/common/Arma 2"}
+ARMA2OASTEAMPATH=${ARMA2OASTEAMPATH:-"/media/s3/games/SteamLibraryProton/steamapps/common/Arma 2 Operation Arrowhead"}
 
 ### delete symlinks
 #find "${SERVERPATH}" -type l -delete
@@ -23,9 +23,9 @@ done
 
 for FILE in $(find ./ -type f); do
 
-	#if [ ! -f "${SERVERPATH}/${FILE,,}" ]; then
+	if [ ! -f "${SERVERPATH}/${FILE,,}" ] || [ -L "${SERVERPATH}/${FILE,,}" ] ; then
 		ln -snf "${ARMA2STEAMPATH}/${FILE}" "${SERVERPATH}/${FILE,,}"
-	#fi
+	fi
 done
 
 ### A2OA
@@ -39,9 +39,9 @@ done
 
 for FILE in $(find ./ -type f ! -path "./ACR/*" ! -path "./PMC/*" ! -path "./BAF/*"); do
 
-	#if [ ! -f "${SERVERPATH}${FILE,,}" ]; then
+	if [ ! -f "${SERVERPATH}${FILE,,}" ] || [ -L "${SERVERPATH}/${FILE,,}" ] ; then
 		ln -snf "${ARMA2OASTEAMPATH}/${FILE}" "${SERVERPATH}/${FILE,,}"
-	#fi
+	fi
 done
